requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.5.2
pypdf2==3.0.1
python-dotenv==1.0.0
psycopg2-binary==2.9.9
requests==2.31.0
pytest==7.4.3
black==23.11.0
isort==5.12.0
mypy==1.7.1
streamlit==1.31.1
streamlit-file-browser==0.1.0 